+++ NEWS.md
@@ -1,4 +1,4 @@
-
+* Added functions to calculate all rebound parameters.
 * Added `sample_eeu_data_path()` function to give the path to an Excel file
   containing same energy efficiency upgrade data.
 * Added a `NEWS.md` file to track changes to the package.